Move buildDerivedName() to unit.ts
